# project-dam-backend/dist/dist/swagger.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
            "mongoose.Types.ObjectId": {
                "type": "string"
            },
...
            "IAddress": {
                "properties": {
                    "street": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zipCode": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "street",
                    "city",
                    "state",
                    "zipCode",
                    "latitude",
                    "longitude"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ISite": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "localName": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/IAddress"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "devicesAtSite": {
                        "items": {
                            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "_id",
                    "localName",
                    "type",
                    "country",
                    "address",
                    "isActive",
                    "createdAt",
                    "updatedAt",
                    "devicesAtSite"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateSiteRequest": {
                "properties": {
                    "localName": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "longitude": {
                                "type": "number",
                                "format": "double"
                            },
                            "latitude": {
                                "type": "number",
                                "format": "double"
                            },
                            "zipCode": {
                                "type": "string"
                            },
                            "state": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "street": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "longitude",
                            "latitude",
                            "zipCode",
                            "state",
                            "city",
                            "street"
                        ],
                        "type": "object"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "localName",
                    "type",
                    "country",
                    "address"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateSiteRequest": {
                "properties": {
                    "localName": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "address": {
                        "properties": {
                            "longitude": {
                                "type": "number",
                                "format": "double"
                            },
                            "latitude": {
                                "type": "number",
                                "format": "double"
                            },
                            "zipCode": {
                                "type": "string"
                            },
                            "state": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            },
                            "street": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "IIp": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "ip": {
                        "type": "string"
                    },
                    "mask": {
                        "type": "number",
                        "format": "double"
                    },
                    "type": {
                        "type": "string"
                    },
                    "logicalEntity": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isActive": {
                        "properties": {
                            "default": {
                                "type": "boolean",
                                "enum": [
                                    true
                                ],
                                "nullable": false
                            },
                            "type": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "default",
                            "type"
                        ],
                        "type": "object"
                    },
                    "devices": {
                        "items": {
                            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "_id",
                    "ip",
                    "mask",
                    "type",
                    "logicalEntity",
                    "createdAt",
                    "updatedAt",
                    "isActive",
                    "devices"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "CreateIpRequest": {
                "properties": {
                    "ip": {
                        "type": "string"
                    },
                    "mask": {
                        "type": "number",
                        "format": "double"
                    },
                    "type": {
                        "type": "string"
                    },
                    "logicalEntity": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "devices": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "ip",
                    "mask",
                    "type",
                    "logicalEntity"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UpdateIpRequest": {
                "properties": {
                    "ip": {
                        "type": "string"
                    },
                    "mask": {
                        "type": "number",
                        "format": "double"
                    },
                    "type": {
                        "type": "string"
                    },
                    "logicalEntity": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "devices": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "project-dam-backend",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
...
        "/sites": {
            "get": {
                "operationId": "GetSites",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pages": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "sites": {
                                            "items": {
                                                "$ref": "#/components/schemas/ISite"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "pages",
                                        "total",
                                        "sites"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Sites"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 1,
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateSite",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ISite"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Sites"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSiteRequest"
                            }
                        }
                    }
                }
            }
        },
        "/sites/{siteId}": {
            "get": {
                "operationId": "GetSite",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ISite"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Sites"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "siteId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateSite",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ISite"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Sites"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "siteId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSiteRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteSite",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Sites"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "siteId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/ips": {
            "get": {
                "operationId": "GetIps",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pages": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "ips": {
                                            "items": {
                                                "$ref": "#/components/schemas/IIp"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "pages",
                                        "total",
                                        "ips"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "IPs"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "default": 1,
                            "format": "double",
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "default": 10,
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            },
            "post": {
                "operationId": "CreateIp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IIp"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "IPs"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIpRequest"
                            }
                        }
                    }
                }
            }
        },
        "/ips/{ipId}": {
            "get": {
                "operationId": "GetIp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IIp"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "IPs"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "ipId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateIp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IIp"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "IPs"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "ipId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIpRequest"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteIp",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ],
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "IPs"
                ],
                "security": [
                    {
                        "jwt": [
                            "admin"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "ipId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}